pyproject.toml
[tool.poetry]
name = "DJANGO-EXAMPLE"
version = "0.1.0"   
description = ""
authors = ["prd-vinh-trinh <vinh.trinh@paradox.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.4"
amqp = "5.2.0"
asgiref = "3.8.1"
asttokens = "2.4.1"
attrs = "24.2.0"
backcall = "0.2.0"
beautifulsoup4 = "4.12.3"
billiard = "4.2.0"
bleach = "6.1.0"
celery = "5.4.0"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "43.0.1"
decorator = "5.1.1"
defusedxml = "0.7.1"
django = "5.1.1"
django-environ = "0.11.2"
django-oauth-toolkit = "3.0.1"
django-redis = "5.4.0"
django-rest-swagger = "2.2.0"
djangorestframework = "3.15.2"
djangorestframework-api-key = "3.0.0"
djangorestframework-simplejwt = "5.3.1"
dnspython = "2.6.1"
docopt = "0.6.2"
drf-nested-forms = "1.1.8"
drf-nested-routers = "0.94.1"
drf-yasg = "1.21.7"
executing = "2.1.0"
fastjsonschema = "2.20.0"
hiredis = "3.0.0"
idna = "3.8"
inflection = "0.5.1"
ipython = "8.12.3"
itypes = "1.2.0"
jedi = "0.19.1"
jinja2 = "3.1.4"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
jupyter-client = "8.6.2"
jupyter-core = "5.7.2"
jupyterlab-pygments = "0.3.0"
jwcrypto = "1.5.6"
kombu = "5.4.1"
markupsafe = "2.1.5"
matplotlib-inline = "0.1.7"
mistune = "3.0.2"
mysqlclient = "2.2.4"
nbclient = "0.10.0"
nbconvert = "7.16.4"
nbformat = "5.10.4"
oauthlib = "3.2.2"
openapi-codec = "1.3.2"
packaging = "24.1"
pandocfilters = "1.5.1"
parso = "0.8.4"
pexpect = "4.9.0"
pickleshare = "0.7.5"
platformdirs = "4.3.2"
prompt-toolkit = "3.0.47"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pycparser = "2.22"
pygments = "2.18.0"
pyjwt = "2.9.0"
pymongo = "4.8.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
pyyaml = "6.0.2"
pyzmq = "26.2.0"
redis = "5.0.8"
referencing = "0.35.1"
requests = "2.32.3"
rpds-py = "0.20.0"
setuptools = "75.1.0"
simplejson = "3.19.3"
six = "1.16.0"
soupsieve = "2.6"
sqlparse = "0.5.1"
stack-data = "0.6.3"
tinycss2 = "1.3.0"
tornado = "6.4.1"
traitlets = "5.14.3"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
vine = "5.1.0"
wcwidth = "0.2.13"
webencodings = "0.5.1"
yarg = "0.1.9"
django-rest-framework-mongoengine = "3.4.1"
mongoengine = "0.29.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
